requirements.txt
transformers>=4.30.0
peft>=0.4.0
accelerate>=0.20.0
datasets>=2.12.0
bitsandbytes>=0.40.0
torch>=2.0.0
torchvision>=0.15.0
pillow>=9.3.0
tqdm>=4.65.0
evaluate>=0.4.0
rouge_score>=0.1.2
nltk>=3.8.0
tensorboard>=2.13.0
sentencepiece>=0.1.99
protobuf>=3.20.3
sentence-transformers>=2.2.2  # Optional, for FGTE distance metric
scipy>=1.10.0  # Optional, for FGTE distance metric
